various fixes for the BSG
svn path=/trunk/boinc/; revision=11127

+++ clientgui/sg_ProjectsComponent.h
@@ -136,7 +136,7 @@ private:
 	wxTimer* checkForMessagesTimer;
 	bool receivedErrorMessage;
 	bool alertMessageDisplayed;
-	static int lastMessageId;
+	static size_t lastMessageId;
 	void CheckForErrorMessages(wxTimerEvent& WXUNUSED(event));
    	void MessagesViewed();
 };
